// src/r/types.ts
export namespace SharedNS {
  export interface IOptions {
    cwd?: string;
  }
}
// ...
export namespace ChangelogNS {
  export interface ICommitAuthorOptions {
    /**
     * Attach author to each changelog commit
     *
     * @default false
     */
    attachAuthor?: boolean;
    /**
     * Type of author name
     *
     * @default 'name'
     */
    authorNameType?: 'name' | 'email';
    /**
     * Get author remote page
     *
     * @param author
     */
    getAuthorPage?(author: ICommitAuthor): string;
  }

  export interface IOptions extends SharedNS.IOptions, ICommitAuthorOptions {
    /**
     * current Version
     *
     * @default version from lerna.json
     */
    version?: string;
    /**
     * Beautify changelog, Remove top-level useless header
     *
     * @see https://github.com/conventional-changelog/conventional-changelog/issues/376
     * @default false
     */
    beautify?: boolean;
    /**
     * Create changelog commit
     *
     * @default false
     */
    commit?: boolean;
    /**
     * Create git push action
     *
     * @default false
     */
    gitPush?: boolean;
  }
  /**
   * Commit author
   */
  export interface ICommitAuthor {
    /**
     * Commit author name
     */
    name: string;
    /**
     * Email name
     */
    emailName: string;
    /**
     * Full email
     */
    email: string;
  }
}